markdown support for operator io label, description, and caption
